+++ azure_sentinel/__init__.py
@@ -1,1 +1,1 @@
-"""Azure Sentinel plugin."""
+"""CLS Microsoft Azure Sentinel plugin."""