+++ azure_sentinel/CHANGELOG.md
@@ -1,14 +1,22 @@
+# 2.0.2
+## Changed
+- Changed tool tip for Alerts Log Type Name, Events Log Type Name and WebTX Log Type Name configuration parameters.
+## Added
+- Added validation for Workspace ID and Primary Key configuration parameters.
+## Updated
+- Updated loggers in plugin.
+
 # 2.0.1
 ## Changed
-- Changed plugin name to Microsoft Azure Sentinel.
+- Changed plugin name from Azure Sentinel to Microsoft Azure Sentinel.
 
 # 2.0.0
 ## Added
-- Added support to sent raw data to the SIEM Platform.
+- Added support to send raw data to the SIEM Platform.
 
 # 1.1.0
 ## Added
-- Support for web transaction logs ingestion.
+- Added support for web transaction logs.
 
 # 1.0.0
 ## Added